Fix stats flicker on filter change
The dashboard stats were originally determining whether to show “Spending Total” or “Spending (filtered)” based on whether the “filters” array is just a list containing the default filter. However, the “filters” array and “dashboardStats” object were being udpated separately. This was causing the “filters” array to finish updating before “dashboardStats”, causing a weird update lag where the “Spending Total” would change to “Spending (filtered)” before the amount updated.

I changed it so that the “filters” that the dashboardStats are referring to are included in the dashboardStats object, so that the amount and title are updated at the same time.