Fixed bug where selected transactions persisted after bulk action
Was mistakenly updating the global transaction list with the copy of the list that has the “selected” flag. This would make is so that whenever you leave the page and come back, all the selected transactions were back. Update the page to only update the local copy of transactions. Then, whenver the component unmounts, the global list gets updated with a fresh copy from the database.